requirements.txt
torch
torchvision
torchaudio
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
PyYAML>=6.0
Pillow>=10.0

# coremltools>=7.0 ; platform_system == "Darwin"
# onnx>=1.15
# onnxruntime>=1.18